requirements.txt
flask
langchain-ibm
langchain-core
pydantic
python-dotenv